Reorganise passport service
