+++ application/services/passport/src/authentication/authentication.module.ts
@@ -1,6 +1,6 @@
 import {Module} from '@nestjs/common'
 
-import {AuthenticationController} from '@/authentication/authentication.controller'
+import {AuthenticationController} from '@/authentication/controllers/authentication/authentication.controller'
 import {AuthenticationService} from '@/authentication/services/authentication/authentication.service'
 import {PasswordHashService} from '@/authentication/services/password-hash/password-hash.service'
 import {StorageModule} from '@/storage/storage.module'
