+++ application/services/passport/src/authentication/services/authentication/authentication-mock.service.ts
@@ -5,6 +5,12 @@ import {FAKE_TOKEN} from '@looport/nest-auth'
 import {AuthenticationService} from '@/authentication/services/authentication/authentication.service'
 import {userMock} from '@/storage/repositories/user/user-mock.repository'
 
+export const validCredentials = {
+  email: 'elliot@ecorp.com',
+  password: 'oj(32$#j',
+}
+export const invalidCredentials = {email: 'elliot', password: 'oj(3'}
+
 type AuthenticationMockService = {
   [method in keyof AuthenticationService]: ReturnType<(typeof mock)['fn']>
 }
