--- application/services/passport/src/authentication/test/authentication.mock.ts
@@ -1,5 +1,0 @@
-export const validCredentials = {
-  email: 'elliot@ecorp.com',
-  password: 'oj(32$#j',
-}
-export const invalidCredentials = {email: 'elliot', password: 'oj(3'}
