+++ application/services/passport/src/storage/repositories/user/user.repository.ts
@@ -2,7 +2,7 @@ import {Injectable} from '@nestjs/common'
 import {InjectRepository} from '@nestjs/typeorm'
 import {Repository} from 'typeorm'
 
-import {UserEntity} from '@/storage/entities/user/user.entity'
+import {UserEntity} from '@/storage/entities/user.entity'
 
 @Injectable()
 export class UserRepository {
