+++ application/services/passport/src/test/authentication/authentication.e2e-test.ts
@@ -10,14 +10,14 @@ import request from 'supertest'
 
 import {AppModule} from '@/app/app.module'
 import {
+  invalidCredentials,
+  validCredentials,
+} from '@/authentication/services/authentication/authentication-mock.service'
+import {
   EMAIL_ALREADY_EXISTS_MESSAGE,
   EMAIL_FIELD_KEY,
   INVALID_LOGIN_CREDENTIALS_MESSAGE,
 } from '@/authentication/services/authentication/constants'
-import {
-  invalidCredentials,
-  validCredentials,
-} from '@/authentication/test/authentication.mock'
 import {UserRepository} from '@/storage/repositories/user/user.repository'
 
 describe('AuthenticationController (e2e)', () => {
