+++ application/services/passport/src/user/controllers/user/user.controller.ts
@@ -3,7 +3,7 @@ import {Serialize} from '@looport/nest-common'
 import {Controller, Get} from '@nestjs/common'
 
 import {UserRepository} from '@/storage/repositories/user/user.repository'
-import {UserDto} from '@/user/dtos/user/user.dto'
+import {UserDto} from '@/user/dto/user/user.dto'
 
 @Serialize(UserDto)
 @Controller('user')
