+++ application/services/passport/src/user/user.module.ts
@@ -3,7 +3,7 @@ import {Module} from '@nestjs/common'
 
 import {StorageModule} from '@/storage/storage.module'
 
-import {UserController} from './user.controller'
+import {UserController} from './controllers/user/user.controller'
 
 @Module({
   controllers: [UserController],
